requirements.txt
TgCrypto
python-telegram-bot
dadjokes
requests